requirements.txt
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
SQLAlchemy==1.4.46
python-dotenv
psycopg2-binary
alembic==1.11.1
DateTime==5.0
gunicorn==20.1.0
Mako==1.2.4
requests==2.28.1
